dom/html/test/test_bug347174.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=347174
-->
<head>
  <title>Test for Bug 347174</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=347174">Mozilla Bug 347174</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 347174 **/
// simple test of readyState during loading, DOMContentLoaded, and complete
// this test passes in IE7
window.readyStateText = [];
window.readyStateText.push("script tag: " + document.readyState);
is(document.readyState, "loading", "document.readyState should be 'loading' when scripts runs initially");

function attachCustomEventListener(element, eventName, command) {
    if (window.addEventListener && !window.opera)
        element.addEventListener(eventName, command, true);
    else if (window.attachEvent)
        element.attachEvent("on" + eventName, command);
}

function showMessage(msg) {
    window.readyStateText.push(msg);
    document.getElementById("display").innerHTML = readyStateText.join("<br>");
}

function load() {
  is(document.readyState, "complete", "document.readyState should be 'complete' on load");
  showMessage("load: " + document.readyState);
  SimpleTest.finish();
}

function readyStateChange() {
    showMessage("readyStateChange: " + document.readyState);
}

function DOMContentLoaded() {
  is(document.readyState, "interactive", "document.readyState should be 'interactive' on DOMContentLoaded");
  showMessage("DOMContentLoaded: " + document.readyState);
}

window.onload=load;

attachCustomEventListener(document, "readystatechange", readyStateChange);
attachCustomEventListener(document, "DOMContentLoaded", DOMContentLoaded);

SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
